teamsdirectrouting-service/requirements.txt
flask
requests
flask-session
PyJWT
pandas
openpyxl
xlrd
